requirements.txt
streamlit>=1.28.0
plotly>=5.15.0
python-dateutil>=2.8.2
google-generativeai>=0.3.0
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0